[LOG] Modulo de menú, modulo de inventario y de proveedores.

+++ ms-camarero-menu/src/main/java/com/es/tfm/ms_camarero_menu/repository/ProductoRepository.java
@@ -2,9 +2,10 @@ package com.es.tfm.ms_camarero_menu.repository;
 
 import com.es.tfm.ms_camarero_menu.model.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-public interface ProductoRepository extends JpaRepository<Producto, Long> {
-    List<Producto> findByBarId(Long barId);
+public interface ProductoRepository extends JpaRepository<Producto, Integer> {
+    List<Producto> findByBarId(Integer barId);
 }
